rename d3d-utils
